♻️ refactor: SignUpModal 상태 전역 상태로 관리 - SignUpModal context 작성 - SignUpModal 스토리북 수정
