+++ src/components/common/Modal/SignUpModal/index.ts
@@ -1,0 +1,2 @@
+export * from "./context";
+export * from "./SignUpModal";
