+++ src/components/common/Modal/index.ts
@@ -2,3 +2,4 @@ export * from "./Modal";
 export * from "./ProfileModal";
 export * from "./SignOutModal";
 export * from "./SignUpModal";
+export * from "./types";
