+++ src/components/common/Modal/types.ts
@@ -1,0 +1,3 @@
+import type { useModal } from "@/application/hooks";
+
+export type ModalProps = ReturnType<typeof useModal>;
