+++ src/components/common/Navigation/Profile/Profile.tsx
@@ -1,11 +1,11 @@
-import { useAuth, useModal } from "@/application/hooks";
+import { useAuth } from "@/application/hooks";
 import { Button } from "@/components/common/Button";
 import { Icon } from "@/components/common/Icon";
-import { ProfileModal, SignUpModal } from "@/components/common/Modal";
+import { ProfileModal, useSignUpModalContext } from "@/components/common/Modal";
 
 export const Profile = () => {
   const { isLoading, isLogin } = useAuth();
-  const modalProps = useModal();
+  const modalProps = useSignUpModalContext();
 
   if (isLoading) return null;
   if (isLogin) return <ProfileModal />;
@@ -15,7 +15,6 @@ export const Profile = () => {
       <Button className="h-32 w-32" onClick={modalProps.onOpen}>
         <Icon name="notloginprofile" />
       </Button>
-      <SignUpModal {...modalProps} />
     </>
   );
 };
